refactor: 위치 잡아주는 로직 컴포넌트에서 분리
