+++ frontend/techpick/src/hooks/index.ts
@@ -7,3 +7,4 @@ export { useCalculateCommandListHeight } from './useCalculateCommandListHeight';
 export { useResetPickFocusOnOutsideClick } from './useResetPickFocusOnOutsideClick';
 export { useClearSelectedPickIdsOnMount } from './useClearSelectedPickIdsOnMount';
 export { useFetchTagList } from './useFetchTagList';
+export { useGetDragOverStyle } from './useGetDragOverStyle';
